+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog for *Blade* project
 ## blade-graphics-0.8.2 (TBD)
 
 - metal: enable fast math, skip debug groups in production
+- add `ComputeCommandEncoder::barrier()` for inline compute-to-compute synchronization within a pass
 
 ## blade-graphics-0.8.1 (28 Mar 2026)